Refactored code to reduce code smells
