+++ src/main/java/org/b3log/solo/SoloServletListener.java
@@ -44,6 +44,7 @@ import org.b3log.solo.model.Option;
 import org.b3log.solo.processor.InitCheckHandler;
 import org.b3log.solo.processor.KanBanNiangProcessor;
 import org.b3log.solo.processor.PermalinkHandler;
+import org.b3log.solo.processor.RouteProcessor;
 import org.b3log.solo.processor.console.*;
 import org.b3log.solo.repository.OptionRepository;
 import org.b3log.solo.service.*;
@@ -58,6 +59,8 @@ import javax.servlet.ServletRequestEvent;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSessionEvent;
 import java.util.Locale;
+
+import static org.b3log.solo.processor.RouteProcessor.routeConsoleProcessors;
 
 /**
  * Solo Servlet listener.
@@ -104,7 +107,9 @@ public final class SoloServletListener extends AbstractServletListener {
         DispatcherServlet.HANDLERS.add(1, new PermalinkHandler());
 
         beanManager = BeanManager.getInstance();
-        routeConsoleProcessors();
+        RouteProcessor routeProcessor = new RouteProcessor();
+        routeProcessor.routeConsoleProcessors();
+
         Stopwatchs.start("Context Initialized");
 
         validateSkin();
@@ -141,30 +146,7 @@ public final class SoloServletListener extends AbstractServletListener {
         final PluginManager pluginManager = beanManager.getReference(PluginManager.class);
         pluginManager.load();
 
-        String header = "" +
-                "████████████████████████████████████████████████████████████████████████\n" +
-                "█                                      █                               █\n" +
-                "█  ██████╗  ██████╗ ██╗      ██████╗   █                               █\n" +
-                "█  ██╔══██╗██╔═══██╗██║     ██╔═══██╗  █  Welcome to Bolo :)           █\n" +
-                "█  ██████╔╝██║   ██║██║     ██║   ██║  █                               █\n" +
-                "█  ██╔══██╗██║   ██║██║     ██║   ██║  █  github.com/adlered/bolo-solo █\n" +
-                "█  ██████╔╝╚██████╔╝███████╗╚██████╔╝  █  Current Version: " + BOLO_VERSION_EN + " █\n" +
-                "█  ╚═════╝  ╚═════╝ ╚══════╝ ╚═════╝   █                               █\n" +
-                "█                                      █                               █\n" +
-                "████✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩████\n" +
-                "█               THANK YOU FOR YOUR CONTRIBUTION TO BOLO !              █\n" +
-                "████✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩████\n" +
-                "█     adlered    (Author)         █    https://github.com/adlered      █\n" +
-                "█     expoli     (Contributor)    █    https://github.com/expoli       █\n" +
-                "█     zeekling   (Contributor)    █    https://github.com/zeekling     █\n" +
-                "█     csfwff     (Contributor)    █    https://github.com/csfwff       █\n" +
-                "█     teahouse   (Contributor)    █    https://github.com/teahouse15   █\n" +
-                "████████████████████████████████████████████████████████████████████████\n" +
-                " \n" +
-                "┌\n" +
-                "├　　HTTP Server Running On:　" + Latkes.getServePath() + "\n" +
-                "├　　JVM Memory:　" + (Runtime.getRuntime().maxMemory() - Runtime.getRuntime().freeMemory()) / 1024 / 1024 + "MB / " + Runtime.getRuntime().maxMemory() / 1024 / 1024 + "MB\n" +
-                "└";
+        String header = getHeader();
         System.out.println("");
         String[] lines = header.split("\n");
         for (String line : lines) {
@@ -194,7 +176,33 @@ public final class SoloServletListener extends AbstractServletListener {
         } catch (JSONException ignored) {
         }
     }
-
+    public String getHeader(){
+        String header = "" +
+                "████████████████████████████████████████████████████████████████████████\n" +
+                "█                                      █                               █\n" +
+                "█  ██████╗  ██████╗ ██╗      ██████╗   █                               █\n" +
+                "█  ██╔══██╗██╔═══██╗██║     ██╔═══██╗  █  Welcome to Bolo :)           █\n" +
+                "█  ██████╔╝██║   ██║██║     ██║   ██║  █                               █\n" +
+                "█  ██╔══██╗██║   ██║██║     ██║   ██║  █  github.com/adlered/bolo-solo █\n" +
+                "█  ██████╔╝╚██████╔╝███████╗╚██████╔╝  █  Current Version: " + BOLO_VERSION_EN + " █\n" +
+                "█  ╚═════╝  ╚═════╝ ╚══════╝ ╚═════╝   █                               █\n" +
+                "█                                      █                               █\n" +
+                "████✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩████\n" +
+                "█               THANK YOU FOR YOUR CONTRIBUTION TO BOLO !              █\n" +
+                "████✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩✩████\n" +
+                "█     adlered    (Author)         █    https://github.com/adlered      █\n" +
+                "█     expoli     (Contributor)    █    https://github.com/expoli       █\n" +
+                "█     zeekling   (Contributor)    █    https://github.com/zeekling     █\n" +
+                "█     csfwff     (Contributor)    █    https://github.com/csfwff       █\n" +
+                "█     teahouse   (Contributor)    █    https://github.com/teahouse15   █\n" +
+                "████████████████████████████████████████████████████████████████████████\n" +
+                " \n" +
+                "┌\n" +
+                "├　　HTTP Server Running On:　" + Latkes.getServePath() + "\n" +
+                "├　　JVM Memory:　" + (Runtime.getRuntime().maxMemory() - Runtime.getRuntime().freeMemory()) / 1024 / 1024 + "MB / " + Runtime.getRuntime().maxMemory() / 1024 / 1024 + "MB\n" +
+                "└";
+        return header;
+    }
     @Override
     public void contextDestroyed(final ServletContextEvent servletContextEvent) {
         super.contextDestroyed(servletContextEvent);
@@ -337,25 +345,56 @@ public final class SoloServletListener extends AbstractServletListener {
     private static void fillBotAttrs(final HttpServletRequest request) {
         final String userAgentStr = request.getHeader("User-Agent");
         final UserAgent userAgent = UserAgent.parseUserAgentString(userAgentStr);
+        BrowserType browserType = determineBrowserType(userAgent,userAgentStr);
+
+
+        request.setAttribute(Keys.HttpRequest.IS_SEARCH_ENGINE_BOT, BrowserType.ROBOT == browserType);
+        request.setAttribute(Keys.HttpRequest.IS_MOBILE_BOT, BrowserType.MOBILE_BROWSER == browserType);
+    }
+
+//    public static BrowserType determineBrowserType(UserAgent userAgent, String userAgentStr) {
+//        BrowserType browserType = userAgent.getBrowser().getBrowserType();
+//
+//        if (StringUtils.containsIgnoreCase(userAgentStr, "mobile")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "MQQBrowser")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "iphone")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "MicroMessenger")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "CFNetwork")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "Android")) {
+//            browserType = BrowserType.MOBILE_BROWSER;
+//        } else if (StringUtils.containsIgnoreCase(userAgentStr, "Iframely")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "Google")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "BUbiNG")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "ltx71")
+//                || StringUtils.containsIgnoreCase(userAgentStr, "py")) {
+//            browserType = BrowserType.ROBOT;
+//        }
+//        return browserType;
+//    }
+
+    public static BrowserType determineBrowserType(UserAgent userAgent, String userAgentStr) {
         BrowserType browserType = userAgent.getBrowser().getBrowserType();
 
-        if (StringUtils.containsIgnoreCase(userAgentStr, "mobile")
+        boolean isMobileBrowser = StringUtils.containsIgnoreCase(userAgentStr, "mobile")
                 || StringUtils.containsIgnoreCase(userAgentStr, "MQQBrowser")
                 || StringUtils.containsIgnoreCase(userAgentStr, "iphone")
                 || StringUtils.containsIgnoreCase(userAgentStr, "MicroMessenger")
                 || StringUtils.containsIgnoreCase(userAgentStr, "CFNetwork")
-                || StringUtils.containsIgnoreCase(userAgentStr, "Android")) {
-            browserType = BrowserType.MOBILE_BROWSER;
-        } else if (StringUtils.containsIgnoreCase(userAgentStr, "Iframely")
+                || StringUtils.containsIgnoreCase(userAgentStr, "Android");
+
+        boolean isRobotBrowser = StringUtils.containsIgnoreCase(userAgentStr, "Iframely")
                 || StringUtils.containsIgnoreCase(userAgentStr, "Google")
                 || StringUtils.containsIgnoreCase(userAgentStr, "BUbiNG")
                 || StringUtils.containsIgnoreCase(userAgentStr, "ltx71")
-                || StringUtils.containsIgnoreCase(userAgentStr, "py")) {
+                || StringUtils.containsIgnoreCase(userAgentStr, "py");
+
+        if (isMobileBrowser) {
+            browserType = BrowserType.MOBILE_BROWSER;
+        } else if (isRobotBrowser) {
             browserType = BrowserType.ROBOT;
         }
 
-        request.setAttribute(Keys.HttpRequest.IS_SEARCH_ENGINE_BOT, BrowserType.ROBOT == browserType);
-        request.setAttribute(Keys.HttpRequest.IS_MOBILE_BOT, BrowserType.MOBILE_BROWSER == browserType);
+        return browserType;
     }
 
     /**
@@ -378,104 +417,104 @@ public final class SoloServletListener extends AbstractServletListener {
     /**
      * 后台控制器使用函数式路由. https://github.com/b3log/solo/issues/12580
      */
-    public static void routeConsoleProcessors() {
-        final BeanManager beanManager = BeanManager.getInstance();
-        final AdminConsole adminConsole = beanManager.getReference(AdminConsole.class);
-        DispatcherServlet.get("/admin-index.do", adminConsole::showAdminIndex);
-        DispatcherServlet.get("/admin-preference.do", adminConsole::showAdminPreferenceFunction);
-        DispatcherServlet.route().get(new String[]{"/admin-article.do",
-                "/admin-article-list.do",
-                "/admin-comment-list.do",
-                "/admin-link-list.do",
-                "/admin-page-list.do",
-                "/admin-others.do",
-                "/admin-draft-list.do",
-                "/admin-user-list.do",
-                "/admin-category-list.do",
-                "/admin-theme-list.do",
-                "/admin-plugin-list.do",
-                "/admin-main.do",
-                "/admin-about.do",
-                "/admin-tool-box.do",
-                "/admin-usite.do"}, adminConsole::showAdminFunctions);
-        DispatcherServlet.get("/console/export/sql", adminConsole::exportSQL);
-        DispatcherServlet.get("/console/export/json", adminConsole::exportJSON);
-        DispatcherServlet.get("/console/export/hexo", adminConsole::exportHexo);
-
-        final ArticleConsole articleConsole = beanManager.getReference(ArticleConsole.class);
-        DispatcherServlet.get("/console/article/push2rhy", articleConsole::pushArticleToCommunity);
-        DispatcherServlet.get("/console/thumbs", articleConsole::getArticleThumbs);
-        DispatcherServlet.get("/console/article/{id}", articleConsole::getArticle);
-        DispatcherServlet.get("/console/articles/status/{status}/{page}/{pageSize}/{windowSize}", articleConsole::getArticles);
-        DispatcherServlet.delete("/console/article/{id}", articleConsole::removeArticle);
-        DispatcherServlet.put("/console/article/unpublish/{id}", articleConsole::cancelPublishArticle);
-        DispatcherServlet.put("/console/article/canceltop/{id}", articleConsole::cancelTopArticle);
-        DispatcherServlet.put("/console/article/puttop/{id}", articleConsole::putTopArticle);
-        DispatcherServlet.put("/console/article/", articleConsole::updateArticle);
-        DispatcherServlet.post("/console/article/", articleConsole::addArticle);
-
-        final CategoryConsole categoryConsole = beanManager.getReference(CategoryConsole.class);
-        DispatcherServlet.put("/console/category/order/", categoryConsole::changeOrder);
-        DispatcherServlet.get("/console/category/{id}", categoryConsole::getCategory);
-        DispatcherServlet.delete("/console/category/{id}", categoryConsole::removeCategory);
-        DispatcherServlet.put("/console/category/", categoryConsole::updateCategory);
-        DispatcherServlet.post("/console/category/", categoryConsole::addCategory);
-        DispatcherServlet.get("/console/categories/{page}/{pageSize}/{windowSize}", categoryConsole::getCategories);
-
-        final CommentConsole commentConsole = beanManager.getReference(CommentConsole.class);
-        DispatcherServlet.delete("/console/article/comment/{id}", commentConsole::removeArticleComment);
-        DispatcherServlet.get("/console/comments/{page}/{pageSize}/{windowSize}", commentConsole::getComments);
-        DispatcherServlet.get("/console/comments/article/{id}", commentConsole::getArticleComments);
-
-        final LinkConsole linkConsole = beanManager.getReference(LinkConsole.class);
-        DispatcherServlet.delete("/console/link/{id}", linkConsole::removeLink);
-        DispatcherServlet.put("/console/link/", linkConsole::updateLink);
-        DispatcherServlet.put("/console/link/order/", linkConsole::changeOrder);
-        DispatcherServlet.post("/console/link/", linkConsole::addLink);
-        DispatcherServlet.get("/console/links/{page}/{pageSize}/{windowSize}", linkConsole::getLinks);
-        DispatcherServlet.get("/console/link/{id}", linkConsole::getLink);
-
-        final PageConsole pageConsole = beanManager.getReference(PageConsole.class);
-        DispatcherServlet.put("/console/page/", pageConsole::updatePage);
-        DispatcherServlet.delete("/console/page/{id}", pageConsole::removePage);
-        DispatcherServlet.post("/console/page/", pageConsole::addPage);
-        DispatcherServlet.put("/console/page/order/", pageConsole::changeOrder);
-        DispatcherServlet.get("/console/page/{id}", pageConsole::getPage);
-        DispatcherServlet.get("/console/pages/{page}/{pageSize}/{windowSize}", pageConsole::getPages);
-
-        final PluginConsole pluginConsole = beanManager.getReference(PluginConsole.class);
-        DispatcherServlet.put("/console/plugin/status/", pluginConsole::setPluginStatus);
-        DispatcherServlet.get("/console/plugins/{page}/{pageSize}/{windowSize}", pluginConsole::getPlugins);
-        DispatcherServlet.post("/console/plugin/toSetting", pluginConsole::toSetting);
-        DispatcherServlet.post("/console/plugin/updateSetting", pluginConsole::updateSetting);
-
-        final PreferenceConsole preferenceConsole = beanManager.getReference(PreferenceConsole.class);
-        DispatcherServlet.get("/console/signs/", preferenceConsole::getSigns);
-        DispatcherServlet.get("/console/preference/", preferenceConsole::getPreference);
-        DispatcherServlet.put("/console/preference/", preferenceConsole::updatePreference);
-
-        final SkinConsole skinConsole = beanManager.getReference(SkinConsole.class);
-        DispatcherServlet.get("/console/skin", skinConsole::getSkin);
-        DispatcherServlet.put("/console/skin", skinConsole::updateSkin);
-
-        final RepairConsole repairConsole = beanManager.getReference(RepairConsole.class);
-        DispatcherServlet.get("/fix/restore-signs", repairConsole::restoreSigns);
-
-        final TagConsole tagConsole = beanManager.getReference(TagConsole.class);
-        DispatcherServlet.get("/console/tags", tagConsole::getTags);
-        DispatcherServlet.get("/console/tag/unused", tagConsole::getUnusedTags);
-
-        final OtherConsole otherConsole = beanManager.getReference(OtherConsole.class);
-        DispatcherServlet.delete("/console/archive/unused", otherConsole::removeUnusedArchives);
-        DispatcherServlet.delete("/console/tag/unused", otherConsole::removeUnusedTags);
-
-        final UserConsole userConsole = beanManager.getReference(UserConsole.class);
-        DispatcherServlet.put("/console/user/", userConsole::updateUser);
-        DispatcherServlet.delete("/console/user/{id}", userConsole::removeUser);
-        DispatcherServlet.get("/console/users/{page}/{pageSize}/{windowSize}", userConsole::getUsers);
-        DispatcherServlet.get("/console/user/{id}", userConsole::getUser);
-        DispatcherServlet.get("/console/changeRole/{id}", userConsole::changeUserRole);
-
-        DispatcherServlet.mapping();
-    }
+//    public static void routeConsoleProcessors() {
+//        final BeanManager beanManager = BeanManager.getInstance();
+//        final AdminConsole adminConsole = beanManager.getReference(AdminConsole.class);
+//        DispatcherServlet.get("/admin-index.do", adminConsole::showAdminIndex);
+//        DispatcherServlet.get("/admin-preference.do", adminConsole::showAdminPreferenceFunction);
+//        DispatcherServlet.route().get(new String[]{"/admin-article.do",
+//                "/admin-article-list.do",
+//                "/admin-comment-list.do",
+//                "/admin-link-list.do",
+//                "/admin-page-list.do",
+//                "/admin-others.do",
+//                "/admin-draft-list.do",
+//                "/admin-user-list.do",
+//                "/admin-category-list.do",
+//                "/admin-theme-list.do",
+//                "/admin-plugin-list.do",
+//                "/admin-main.do",
+//                "/admin-about.do",
+//                "/admin-tool-box.do",
+//                "/admin-usite.do"}, adminConsole::showAdminFunctions);
+//        DispatcherServlet.get("/console/export/sql", adminConsole::exportSQL);
+//        DispatcherServlet.get("/console/export/json", adminConsole::exportJSON);
+//        DispatcherServlet.get("/console/export/hexo", adminConsole::exportHexo);
+//
+//        final ArticleConsole articleConsole = beanManager.getReference(ArticleConsole.class);
+//        DispatcherServlet.get("/console/article/push2rhy", articleConsole::pushArticleToCommunity);
+//        DispatcherServlet.get("/console/thumbs", articleConsole::getArticleThumbs);
+//        DispatcherServlet.get("/console/article/{id}", articleConsole::getArticle);
+//        DispatcherServlet.get("/console/articles/status/{status}/{page}/{pageSize}/{windowSize}", articleConsole::getArticles);
+//        DispatcherServlet.delete("/console/article/{id}", articleConsole::removeArticle);
+//        DispatcherServlet.put("/console/article/unpublish/{id}", articleConsole::cancelPublishArticle);
+//        DispatcherServlet.put("/console/article/canceltop/{id}", articleConsole::cancelTopArticle);
+//        DispatcherServlet.put("/console/article/puttop/{id}", articleConsole::putTopArticle);
+//        DispatcherServlet.put("/console/article/", articleConsole::updateArticle);
+//        DispatcherServlet.post("/console/article/", articleConsole::addArticle);
+//
+//        final CategoryConsole categoryConsole = beanManager.getReference(CategoryConsole.class);
+//        DispatcherServlet.put("/console/category/order/", categoryConsole::changeOrder);
+//        DispatcherServlet.get("/console/category/{id}", categoryConsole::getCategory);
+//        DispatcherServlet.delete("/console/category/{id}", categoryConsole::removeCategory);
+//        DispatcherServlet.put("/console/category/", categoryConsole::updateCategory);
+//        DispatcherServlet.post("/console/category/", categoryConsole::addCategory);
+//        DispatcherServlet.get("/console/categories/{page}/{pageSize}/{windowSize}", categoryConsole::getCategories);
+//
+//        final CommentConsole commentConsole = beanManager.getReference(CommentConsole.class);
+//        DispatcherServlet.delete("/console/article/comment/{id}", commentConsole::removeArticleComment);
+//        DispatcherServlet.get("/console/comments/{page}/{pageSize}/{windowSize}", commentConsole::getComments);
+//        DispatcherServlet.get("/console/comments/article/{id}", commentConsole::getArticleComments);
+//
+//        final LinkConsole linkConsole = beanManager.getReference(LinkConsole.class);
+//        DispatcherServlet.delete("/console/link/{id}", linkConsole::removeLink);
+//        DispatcherServlet.put("/console/link/", linkConsole::updateLink);
+//        DispatcherServlet.put("/console/link/order/", linkConsole::changeOrder);
+//        DispatcherServlet.post("/console/link/", linkConsole::addLink);
+//        DispatcherServlet.get("/console/links/{page}/{pageSize}/{windowSize}", linkConsole::getLinks);
+//        DispatcherServlet.get("/console/link/{id}", linkConsole::getLink);
+//
+//        final PageConsole pageConsole = beanManager.getReference(PageConsole.class);
+//        DispatcherServlet.put("/console/page/", pageConsole::updatePage);
+//        DispatcherServlet.delete("/console/page/{id}", pageConsole::removePage);
+//        DispatcherServlet.post("/console/page/", pageConsole::addPage);
+//        DispatcherServlet.put("/console/page/order/", pageConsole::changeOrder);
+//        DispatcherServlet.get("/console/page/{id}", pageConsole::getPage);
+//        DispatcherServlet.get("/console/pages/{page}/{pageSize}/{windowSize}", pageConsole::getPages);
+//
+//        final PluginConsole pluginConsole = beanManager.getReference(PluginConsole.class);
+//        DispatcherServlet.put("/console/plugin/status/", pluginConsole::setPluginStatus);
+//        DispatcherServlet.get("/console/plugins/{page}/{pageSize}/{windowSize}", pluginConsole::getPlugins);
+//        DispatcherServlet.post("/console/plugin/toSetting", pluginConsole::toSetting);
+//        DispatcherServlet.post("/console/plugin/updateSetting", pluginConsole::updateSetting);
+//
+//        final PreferenceConsole preferenceConsole = beanManager.getReference(PreferenceConsole.class);
+//        DispatcherServlet.get("/console/signs/", preferenceConsole::getSigns);
+//        DispatcherServlet.get("/console/preference/", preferenceConsole::getPreference);
+//        DispatcherServlet.put("/console/preference/", preferenceConsole::updatePreference);
+//
+//        final SkinConsole skinConsole = beanManager.getReference(SkinConsole.class);
+//        DispatcherServlet.get("/console/skin", skinConsole::getSkin);
+//        DispatcherServlet.put("/console/skin", skinConsole::updateSkin);
+//
+//        final RepairConsole repairConsole = beanManager.getReference(RepairConsole.class);
+//        DispatcherServlet.get("/fix/restore-signs", repairConsole::restoreSigns);
+//
+//        final TagConsole tagConsole = beanManager.getReference(TagConsole.class);
+//        DispatcherServlet.get("/console/tags", tagConsole::getTags);
+//        DispatcherServlet.get("/console/tag/unused", tagConsole::getUnusedTags);
+//
+//        final OtherConsole otherConsole = beanManager.getReference(OtherConsole.class);
+//        DispatcherServlet.delete("/console/archive/unused", otherConsole::removeUnusedArchives);
+//        DispatcherServlet.delete("/console/tag/unused", otherConsole::removeUnusedTags);
+//
+//        final UserConsole userConsole = beanManager.getReference(UserConsole.class);
+//        DispatcherServlet.put("/console/user/", userConsole::updateUser);
+//        DispatcherServlet.delete("/console/user/{id}", userConsole::removeUser);
+//        DispatcherServlet.get("/console/users/{page}/{pageSize}/{windowSize}", userConsole::getUsers);
+//        DispatcherServlet.get("/console/user/{id}", userConsole::getUser);
+//        DispatcherServlet.get("/console/changeRole/{id}", userConsole::changeUserRole);
+//
+//        DispatcherServlet.mapping();
+//    }
 }
